requirements-analysis.txt
bertopic>=0.16.0
umap-learn>=0.5.5
hdbscan>=0.8.33
sentence-transformers>=3.0.0
scikit-learn>=1.3.0
transformers>=4.40.0
torch>=2.2.0
optimum[onnxruntime]>=1.19.0
plotly>=5.18.0
dash>=2.17.0
dash-bootstrap-components>=1.5.0
dash-cytoscape>=1.0.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
safetensors>=0.4.0
